pyproject.toml
[project]
name = "paternologia"
version = "0.1.0"
description = "Aplikacja do zarządzania konfiguracjami MIDI dla utworów muzycznych"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "hatchling>=1.28.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paternologia"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
